+++ tsup.config.ts
@@ -58,7 +58,6 @@ export default defineConfig({
   esbuildOptions(options) {
     options.alias = {
       '@/lib/utils': './lib/utils.ts',
-      '@/lib/dropdown-placement': './lib/dropdown-placement.ts',
       '@/lib/floating-dropdown': './lib/floating-dropdown.tsx',
       '@/lib/bypass-modal-scroll-lock': './lib/bypass-modal-scroll-lock.ts',
       '@/lib/use-mobile': './lib/use-mobile.ts',